// packages/schemas/src/enums/notification-type.enum.ts
export enum NotificationTypeEnum {
    TRIAL_EXPIRING = 'TRIAL_EXPIRING',
    TRIAL_EXPIRED = 'TRIAL_EXPIRED',
    PAYMENT_DUE = 'PAYMENT_DUE',
    PAYMENT_FAILED = 'PAYMENT_FAILED',
    PAYMENT_SUCCESS = 'PAYMENT_SUCCESS',
    SUBSCRIPTION_RENEWED = 'SUBSCRIPTION_RENEWED',
    SUBSCRIPTION_CANCELLED = 'SUBSCRIPTION_CANCELLED',
    INVOICE_GENERATED = 'INVOICE_GENERATED',
    SERVICE_ORDER_UPDATE = 'SERVICE_ORDER_UPDATE',
    CAMPAIGN_STATUS_CHANGE = 'CAMPAIGN_STATUS_CHANGE',
    LISTING_APPROVED = 'LISTING_APPROVED',
    LISTING_REJECTED = 'LISTING_REJECTED',
    SYSTEM_MAINTENANCE = 'SYSTEM_MAINTENANCE',
    CUSTOM = 'CUSTOM'
}